split traits of storage & merge rpc implementation
