--- light-client-bin/src/rpc/rpc_rocksdb.rs
@@ -1,704 +1,0 @@
-use jsonrpc_core::{Error, Result};
-
-use ckb_light_client_lib::{
-    service::{
-        Cell, CellType, CellsCapacity, Order, Pagination, ScriptStatus, ScriptType, SearchKey,
-        SetScriptsCommand, Tx, TxWithCell, TxWithCells,
-    },
-    storage::{self, extract_raw_data, Key, KeyPrefix, LAST_STATE_KEY},
-};
-
-use ckb_jsonrpc_types::{JsonBytes, Uint32};
-use ckb_types::{core, packed, prelude::*};
-use rocksdb::{
-    ops::{Get, Iterate},
-    Direction, IteratorMode,
-};
-
-use crate::rpc::{BlockFilterRpc, BlockFilterRpcImpl};
-use ckb_light_client_lib::storage::db::StorageHighLevelOperations;
-
-impl BlockFilterRpc for BlockFilterRpcImpl {
-    fn set_scripts(
-        &self,
-        scripts: Vec<ScriptStatus>,
-        command: Option<SetScriptsCommand>,
-    ) -> Result<()> {
-        let (tx, rx) = std::sync::mpsc::channel();
-        let swc = self.swc.clone();
-        std::thread::spawn(move || {
-            let mut matched_blocks = swc.matched_blocks().blocking_write();
-            let scripts = scripts.into_iter().map(Into::into).collect();
-            swc.storage()
-                .update_filter_scripts(scripts, command.map(Into::into).unwrap_or_default());
-            matched_blocks.clear();
-            tx.send(()).unwrap();
-        });
-        rx.recv().unwrap();
-        Ok(())
-    }
-
-    fn get_scripts(&self) -> Result<Vec<ScriptStatus>> {
-        let scripts = self.swc.storage().get_filter_scripts();
-        Ok(scripts.into_iter().map(Into::into).collect())
-    }
-
-    fn get_cells(
-        &self,
-        search_key: SearchKey,
-        order: Order,
-        limit: Uint32,
-        after_cursor: Option<JsonBytes>,
-    ) -> Result<Pagination<Cell>> {
-        let (prefix, from_key, direction, skip) = build_query_options(
-            &search_key,
-            KeyPrefix::CellLockScript,
-            KeyPrefix::CellTypeScript,
-            order,
-            after_cursor,
-        )?;
-        let limit = limit.value() as usize;
-        if limit == 0 {
-            return Err(Error::invalid_params("limit should be greater than 0"));
-        }
-        let with_data = search_key.with_data.unwrap_or(true);
-        let filter_script_type = match search_key.script_type {
-            ScriptType::Lock => ScriptType::Type,
-            ScriptType::Type => ScriptType::Lock,
-        };
-        let (
-            filter_prefix,
-            filter_script_len_range,
-            filter_output_data_len_range,
-            filter_output_capacity_range,
-            filter_block_range,
-        ) = build_filter_options(search_key)?;
-        let mode = IteratorMode::From(from_key.as_ref(), direction);
-        let snapshot = self.swc.storage().snapshot();
-        let iter = snapshot.iterator(mode).skip(skip);
-
-        let mut last_key = Vec::new();
-        let cells = iter
-            .take_while(|(key, _value)| key.starts_with(&prefix))
-            .filter_map(|(key, value)| {
-                let tx_hash = packed::Byte32::from_slice(&value).expect("stored tx hash");
-                let output_index = u32::from_be_bytes(
-                    key[key.len() - 4..]
-                        .try_into()
-                        .expect("stored output_index"),
-                );
-                let tx_index = u32::from_be_bytes(
-                    key[key.len() - 8..key.len() - 4]
-                        .try_into()
-                        .expect("stored tx_index"),
-                );
-                let block_number = u64::from_be_bytes(
-                    key[key.len() - 16..key.len() - 8]
-                        .try_into()
-                        .expect("stored block_number"),
-                );
-
-                let tx = packed::Transaction::from_slice(
-                    &snapshot
-                        .get(Key::TxHash(&tx_hash).into_vec())
-                        .expect("get tx should be OK")
-                        .expect("stored tx")[12..],
-                )
-                .expect("from stored tx slice should be OK");
-                let output = tx
-                    .raw()
-                    .outputs()
-                    .get(output_index as usize)
-                    .expect("get output by index should be OK");
-                let output_data = tx
-                    .raw()
-                    .outputs_data()
-                    .get(output_index as usize)
-                    .expect("get output data by index should be OK");
-
-                if let Some(prefix) = filter_prefix.as_ref() {
-                    match filter_script_type {
-                        ScriptType::Lock => {
-                            if !extract_raw_data(&output.lock())
-                                .as_slice()
-                                .starts_with(prefix)
-                            {
-                                return None;
-                            }
-                        }
-                        ScriptType::Type => {
-                            if output.type_().is_none()
-                                || !extract_raw_data(&output.type_().to_opt().unwrap())
-                                    .as_slice()
-                                    .starts_with(prefix)
-                            {
-                                return None;
-                            }
-                        }
-                    }
-                }
-
-                if let Some([r0, r1]) = filter_script_len_range {
-                    match filter_script_type {
-                        ScriptType::Lock => {
-                            let script_len = extract_raw_data(&output.lock()).len();
-                            if script_len < r0 || script_len > r1 {
-                                return None;
-                            }
-                        }
-                        ScriptType::Type => {
-                            let script_len = output
-                                .type_()
-                                .to_opt()
-                                .map(|script| extract_raw_data(&script).len())
-                                .unwrap_or_default();
-                            if script_len < r0 || script_len > r1 {
-                                return None;
-                            }
-                        }
-                    }
-                }
-
-                if let Some([r0, r1]) = filter_output_data_len_range {
-                    if output_data.len() < r0 || output_data.len() >= r1 {
-                        return None;
-                    }
-                }
-
-                if let Some([r0, r1]) = filter_output_capacity_range {
-                    let capacity: core::Capacity = output.capacity().unpack();
-                    if capacity < r0 || capacity >= r1 {
-                        return None;
-                    }
-                }
-
-                if let Some([r0, r1]) = filter_block_range {
-                    if block_number < r0 || block_number >= r1 {
-                        return None;
-                    }
-                }
-
-                last_key = key.to_vec();
-
-                Some(Cell {
-                    output: output.into(),
-                    output_data: if with_data {
-                        Some(output_data.into())
-                    } else {
-                        None
-                    },
-                    out_point: packed::OutPoint::new(tx_hash, output_index).into(),
-                    block_number: block_number.into(),
-                    tx_index: tx_index.into(),
-                })
-            })
-            .take(limit)
-            .collect::<Vec<_>>();
-
-        Ok(Pagination {
-            objects: cells,
-            last_cursor: JsonBytes::from_vec(last_key),
-        })
-    }
-
-    fn get_transactions(
-        &self,
-        search_key: SearchKey,
-        order: Order,
-        limit: Uint32,
-        after_cursor: Option<JsonBytes>,
-    ) -> Result<Pagination<Tx>> {
-        let (prefix, from_key, direction, skip) = build_query_options(
-            &search_key,
-            KeyPrefix::TxLockScript,
-            KeyPrefix::TxTypeScript,
-            order,
-            after_cursor,
-        )?;
-        let limit = limit.value() as usize;
-        if limit == 0 {
-            return Err(Error::invalid_params("limit should be greater than 0"));
-        }
-
-        let (filter_script, filter_block_range) = if let Some(filter) = search_key.filter.as_ref() {
-            if filter.output_data_len_range.is_some() {
-                return Err(Error::invalid_params(
-                    "doesn't support search_key.filter.output_data_len_range parameter",
-                ));
-            }
-            if filter.output_capacity_range.is_some() {
-                return Err(Error::invalid_params(
-                    "doesn't support search_key.filter.output_capacity_range parameter",
-                ));
-            }
-            let filter_script: Option<packed::Script> =
-                filter.script.as_ref().map(|script| script.clone().into());
-            let filter_block_range: Option<[core::BlockNumber; 2]> =
-                filter.block_range.map(|r| [r[0].into(), r[1].into()]);
-            (filter_script, filter_block_range)
-        } else {
-            (None, None)
-        };
-
-        let filter_script_type = match search_key.script_type {
-            ScriptType::Lock => ScriptType::Type,
-            ScriptType::Type => ScriptType::Lock,
-        };
-
-        let mode = IteratorMode::From(from_key.as_ref(), direction);
-        let snapshot = self.swc.storage().snapshot();
-        let iter = snapshot.iterator(mode).skip(skip);
-
-        if search_key.group_by_transaction.unwrap_or_default() {
-            let mut tx_with_cells: Vec<TxWithCells> = Vec::new();
-            let mut last_key = Vec::new();
-
-            for (key, value) in iter.take_while(|(key, _value)| key.starts_with(&prefix)) {
-                let tx_hash = packed::Byte32::from_slice(&value).expect("stored tx hash");
-                if tx_with_cells.len() == limit
-                    && tx_with_cells.last_mut().unwrap().transaction.hash != tx_hash.unpack()
-                {
-                    break;
-                }
-                last_key = key.to_vec();
-                let tx = packed::Transaction::from_slice(
-                    &snapshot
-                        .get(Key::TxHash(&tx_hash).into_vec())
-                        .expect("get tx should be OK")
-                        .expect("stored tx")[12..],
-                )
-                .expect("from stored tx slice should be OK");
-
-                let block_number = u64::from_be_bytes(
-                    key[key.len() - 17..key.len() - 9]
-                        .try_into()
-                        .expect("stored block_number"),
-                );
-                let tx_index = u32::from_be_bytes(
-                    key[key.len() - 9..key.len() - 5]
-                        .try_into()
-                        .expect("stored tx_index"),
-                );
-                let io_index = u32::from_be_bytes(
-                    key[key.len() - 5..key.len() - 1]
-                        .try_into()
-                        .expect("stored io_index"),
-                );
-                let io_type = if *key.last().expect("stored io_type") == 0 {
-                    CellType::Input
-                } else {
-                    CellType::Output
-                };
-
-                if let Some(filter_script) = filter_script.as_ref() {
-                    let filter_script_matched = match filter_script_type {
-                        ScriptType::Lock => snapshot
-                            .get(
-                                Key::TxLockScript(
-                                    filter_script,
-                                    block_number,
-                                    tx_index,
-                                    io_index,
-                                    match io_type {
-                                        CellType::Input => storage::CellType::Input,
-                                        CellType::Output => storage::CellType::Output,
-                                    },
-                                )
-                                .into_vec(),
-                            )
-                            .expect("get TxLockScript should be OK")
-                            .is_some(),
-                        ScriptType::Type => snapshot
-                            .get(
-                                Key::TxTypeScript(
-                                    filter_script,
-                                    block_number,
-                                    tx_index,
-                                    io_index,
-                                    match io_type {
-                                        CellType::Input => storage::CellType::Input,
-                                        CellType::Output => storage::CellType::Output,
-                                    },
-                                )
-                                .into_vec(),
-                            )
-                            .expect("get TxTypeScript should be OK")
-                            .is_some(),
-                    };
-
-                    if !filter_script_matched {
-                        continue;
-                    }
-                }
-
-                if let Some([r0, r1]) = filter_block_range {
-                    if block_number < r0 || block_number >= r1 {
-                        continue;
-                    }
-                }
-
-                let last_tx_hash_is_same = tx_with_cells
-                    .last_mut()
-                    .map(|last| {
-                        if last.transaction.hash == tx_hash.unpack() {
-                            last.cells.push((io_type.clone(), io_index.into()));
-                            true
-                        } else {
-                            false
-                        }
-                    })
-                    .unwrap_or_default();
-
-                if !last_tx_hash_is_same {
-                    tx_with_cells.push(TxWithCells {
-                        transaction: tx.into_view().into(),
-                        block_number: block_number.into(),
-                        tx_index: tx_index.into(),
-                        cells: vec![(io_type, io_index.into())],
-                    });
-                }
-            }
-
-            Ok(Pagination {
-                objects: tx_with_cells.into_iter().map(Tx::Grouped).collect(),
-                last_cursor: JsonBytes::from_vec(last_key),
-            })
-        } else {
-            let mut last_key = Vec::new();
-            let txs = iter
-                .take_while(|(key, _value)| key.starts_with(&prefix))
-                .filter_map(|(key, value)| {
-                    let tx_hash = packed::Byte32::from_slice(&value).expect("stored tx hash");
-                    let tx = packed::Transaction::from_slice(
-                        &snapshot
-                            .get(Key::TxHash(&tx_hash).into_vec())
-                            .expect("get tx should be OK")
-                            .expect("stored tx")[12..],
-                    )
-                    .expect("from stored tx slice should be OK");
-
-                    let block_number = u64::from_be_bytes(
-                        key[key.len() - 17..key.len() - 9]
-                            .try_into()
-                            .expect("stored block_number"),
-                    );
-                    let tx_index = u32::from_be_bytes(
-                        key[key.len() - 9..key.len() - 5]
-                            .try_into()
-                            .expect("stored tx_index"),
-                    );
-                    let io_index = u32::from_be_bytes(
-                        key[key.len() - 5..key.len() - 1]
-                            .try_into()
-                            .expect("stored io_index"),
-                    );
-                    let io_type = if *key.last().expect("stored io_type") == 0 {
-                        CellType::Input
-                    } else {
-                        CellType::Output
-                    };
-
-                    if let Some(filter_script) = filter_script.as_ref() {
-                        match filter_script_type {
-                            ScriptType::Lock => {
-                                snapshot
-                                    .get(
-                                        Key::TxLockScript(
-                                            filter_script,
-                                            block_number,
-                                            tx_index,
-                                            io_index,
-                                            match io_type {
-                                                CellType::Input => storage::CellType::Input,
-                                                CellType::Output => storage::CellType::Output,
-                                            },
-                                        )
-                                        .into_vec(),
-                                    )
-                                    .expect("get TxLockScript should be OK")?;
-                            }
-                            ScriptType::Type => {
-                                snapshot
-                                    .get(
-                                        Key::TxTypeScript(
-                                            filter_script,
-                                            block_number,
-                                            tx_index,
-                                            io_index,
-                                            match io_type {
-                                                CellType::Input => storage::CellType::Input,
-                                                CellType::Output => storage::CellType::Output,
-                                            },
-                                        )
-                                        .into_vec(),
-                                    )
-                                    .expect("get TxTypeScript should be OK")?;
-                            }
-                        }
-                    }
-
-                    if let Some([r0, r1]) = filter_block_range {
-                        if block_number < r0 || block_number >= r1 {
-                            return None;
-                        }
-                    }
-
-                    last_key = key.to_vec();
-                    Some(Tx::Ungrouped(TxWithCell {
-                        transaction: tx.into_view().into(),
-                        block_number: block_number.into(),
-                        tx_index: tx_index.into(),
-                        io_index: io_index.into(),
-                        io_type,
-                    }))
-                })
-                .take(limit)
-                .collect::<Vec<_>>();
-
-            Ok(Pagination {
-                objects: txs,
-                last_cursor: JsonBytes::from_vec(last_key),
-            })
-        }
-    }
-
-    fn get_cells_capacity(&self, search_key: SearchKey) -> Result<CellsCapacity> {
-        let (prefix, from_key, direction, skip) = build_query_options(
-            &search_key,
-            KeyPrefix::CellLockScript,
-            KeyPrefix::CellTypeScript,
-            Order::Asc,
-            None,
-        )?;
-        let filter_script_type = match search_key.script_type {
-            ScriptType::Lock => ScriptType::Type,
-            ScriptType::Type => ScriptType::Lock,
-        };
-        let (
-            filter_prefix,
-            filter_script_len_range,
-            filter_output_data_len_range,
-            filter_output_capacity_range,
-            filter_block_range,
-        ) = build_filter_options(search_key)?;
-        let mode = IteratorMode::From(from_key.as_ref(), direction);
-        let snapshot = self.swc.storage().snapshot();
-        let iter = snapshot.iterator(mode).skip(skip);
-
-        let capacity: u64 = iter
-            .take_while(|(key, _value)| key.starts_with(&prefix))
-            .filter_map(|(key, value)| {
-                let tx_hash = packed::Byte32::from_slice(&value).expect("stored tx hash");
-                let output_index = u32::from_be_bytes(
-                    key[key.len() - 4..]
-                        .try_into()
-                        .expect("stored output_index"),
-                );
-                let block_number = u64::from_be_bytes(
-                    key[key.len() - 16..key.len() - 8]
-                        .try_into()
-                        .expect("stored block_number"),
-                );
-
-                let tx = packed::Transaction::from_slice(
-                    &snapshot
-                        .get(Key::TxHash(&tx_hash).into_vec())
-                        .expect("get tx should be OK")
-                        .expect("stored tx")[12..],
-                )
-                .expect("from stored tx slice should be OK");
-                let output = tx
-                    .raw()
-                    .outputs()
-                    .get(output_index as usize)
-                    .expect("get output by index should be OK");
-                let output_data = tx
-                    .raw()
-                    .outputs_data()
-                    .get(output_index as usize)
-                    .expect("get output data by index should be OK");
-
-                if let Some(prefix) = filter_prefix.as_ref() {
-                    match filter_script_type {
-                        ScriptType::Lock => {
-                            if !extract_raw_data(&output.lock())
-                                .as_slice()
-                                .starts_with(prefix)
-                            {
-                                return None;
-                            }
-                        }
-                        ScriptType::Type => {
-                            if output.type_().is_none()
-                                || !extract_raw_data(&output.type_().to_opt().unwrap())
-                                    .as_slice()
-                                    .starts_with(prefix)
-                            {
-                                return None;
-                            }
-                        }
-                    }
-                }
-
-                if let Some([r0, r1]) = filter_script_len_range {
-                    match filter_script_type {
-                        ScriptType::Lock => {
-                            let script_len = extract_raw_data(&output.lock()).len();
-                            if script_len < r0 || script_len > r1 {
-                                return None;
-                            }
-                        }
-                        ScriptType::Type => {
-                            let script_len = output
-                                .type_()
-                                .to_opt()
-                                .map(|script| extract_raw_data(&script).len())
-                                .unwrap_or_default();
-                            if script_len < r0 || script_len > r1 {
-                                return None;
-                            }
-                        }
-                    }
-                }
-
-                if let Some([r0, r1]) = filter_output_data_len_range {
-                    if output_data.len() < r0 || output_data.len() >= r1 {
-                        return None;
-                    }
-                }
-
-                if let Some([r0, r1]) = filter_output_capacity_range {
-                    let capacity: core::Capacity = output.capacity().unpack();
-                    if capacity < r0 || capacity >= r1 {
-                        return None;
-                    }
-                }
-
-                if let Some([r0, r1]) = filter_block_range {
-                    if block_number < r0 || block_number >= r1 {
-                        return None;
-                    }
-                }
-
-                Some(Unpack::<core::Capacity>::unpack(&output.capacity()).as_u64())
-            })
-            .sum();
-
-        let key = Key::Meta(LAST_STATE_KEY).into_vec();
-        let tip_header = snapshot
-            .get(key)
-            .expect("snapshot get last state should be ok")
-            .map(|data| packed::HeaderReader::from_slice_should_be_ok(&data[32..]).to_entity())
-            .expect("tip header should be inited");
-        Ok(CellsCapacity {
-            capacity: capacity.into(),
-            block_hash: tip_header.calc_header_hash().unpack(),
-            block_number: tip_header.raw().number().unpack(),
-        })
-    }
-}
-
-const MAX_PREFIX_SEARCH_SIZE: usize = u16::MAX as usize;
-
-// a helper fn to build query options from search paramters, returns prefix, from_key, direction and skip offset
-pub fn build_query_options(
-    search_key: &SearchKey,
-    lock_prefix: KeyPrefix,
-    type_prefix: KeyPrefix,
-    order: Order,
-    after_cursor: Option<JsonBytes>,
-) -> Result<(Vec<u8>, Vec<u8>, Direction, usize)> {
-    let mut prefix = match search_key.script_type {
-        ScriptType::Lock => vec![lock_prefix as u8],
-        ScriptType::Type => vec![type_prefix as u8],
-    };
-    let script: packed::Script = search_key.script.clone().into();
-    let args_len = script.args().len();
-    if args_len > MAX_PREFIX_SEARCH_SIZE {
-        return Err(Error::invalid_params(format!(
-            "search_key.script.args len should be less than {}",
-            MAX_PREFIX_SEARCH_SIZE
-        )));
-    }
-    prefix.extend_from_slice(extract_raw_data(&script).as_slice());
-
-    let (from_key, direction, skip) = match order {
-        Order::Asc => after_cursor.map_or_else(
-            || (prefix.clone(), Direction::Forward, 0),
-            |json_bytes| (json_bytes.as_bytes().into(), Direction::Forward, 1),
-        ),
-        Order::Desc => after_cursor.map_or_else(
-            || {
-                (
-                    [
-                        prefix.clone(),
-                        vec![0xff; MAX_PREFIX_SEARCH_SIZE - args_len],
-                    ]
-                    .concat(),
-                    Direction::Reverse,
-                    0,
-                )
-            },
-            |json_bytes| (json_bytes.as_bytes().into(), Direction::Reverse, 1),
-        ),
-    };
-
-    Ok((prefix, from_key, direction, skip))
-}
-
-// a helper fn to build filter options from search paramters, returns prefix, output_data_len_range, output_capacity_range and block_range
-#[allow(clippy::type_complexity)]
-pub fn build_filter_options(
-    search_key: SearchKey,
-) -> Result<(
-    Option<Vec<u8>>,
-    Option<[usize; 2]>,
-    Option<[usize; 2]>,
-    Option<[core::Capacity; 2]>,
-    Option<[core::BlockNumber; 2]>,
-)> {
-    let filter = search_key.filter.unwrap_or_default();
-    let filter_script_prefix = if let Some(script) = filter.script {
-        let script: packed::Script = script.into();
-        if script.args().len() > MAX_PREFIX_SEARCH_SIZE {
-            return Err(Error::invalid_params(format!(
-                "search_key.filter.script.args len should be less than {}",
-                MAX_PREFIX_SEARCH_SIZE
-            )));
-        }
-        let mut prefix = Vec::new();
-        prefix.extend_from_slice(extract_raw_data(&script).as_slice());
-        Some(prefix)
-    } else {
-        None
-    };
-
-    let filter_script_len_range = filter.script_len_range.map(|[r0, r1]| {
-        [
-            Into::<u64>::into(r0) as usize,
-            Into::<u64>::into(r1) as usize,
-        ]
-    });
-
-    let filter_output_data_len_range = filter.output_data_len_range.map(|[r0, r1]| {
-        [
-            Into::<u64>::into(r0) as usize,
-            Into::<u64>::into(r1) as usize,
-        ]
-    });
-    let filter_output_capacity_range = filter.output_capacity_range.map(|[r0, r1]| {
-        [
-            core::Capacity::shannons(r0.into()),
-            core::Capacity::shannons(r1.into()),
-        ]
-    });
-    let filter_block_range = filter.block_range.map(|r| [r[0].into(), r[1].into()]);
-
-    Ok((
-        filter_script_prefix,
-        filter_script_len_range,
-        filter_output_data_len_range,
-        filter_output_capacity_range,
-        filter_block_range,
-    ))
-}
