+++ light-client-lib/src/protocols/filter/block_filter.rs
@@ -1,6 +1,6 @@
 use super::{components, BAD_MESSAGE_BAN_TIME};
 use crate::protocols::{Peers, Status, StatusCode};
-use crate::storage::db::StorageHighLevelOperations;
+use crate::storage::db::{StorageGeneralOperations, StorageGetPinnedRelatedOperations, StorageHighLevelOperations};
 use crate::storage::Storage;
 use crate::types::{Duration, Instant, RwLock};
 use crate::utils::network::prove_or_download_matched_blocks;
