+++ light-client-lib/src/protocols/filter/components/block_filters_process.rs
@@ -1,6 +1,6 @@
 use crate::protocols::FilterProtocol;
 use crate::protocols::{Status, StatusCode};
-use crate::storage::db::StorageHighLevelOperations;
+use crate::storage::db::{StorageGeneralOperations, StorageGetPinnedRelatedOperations, StorageHighLevelOperations};
 use crate::utils::network::prove_or_download_matched_blocks;
 use ckb_constant::sync::INIT_BLOCKS_IN_TRANSIT_PER_PEER;
 use ckb_network::{BoxedCKBProtocolContext, PeerIndex};
