+++ light-client-lib/src/protocols/light_client/components/send_transactions_proof.rs
@@ -6,7 +6,7 @@ use ckb_types::{
 };
 use log::{debug, error};
 
-use crate::{protocols::light_client::components::verify_extra_hash, storage::HeaderWithExtension};
+use crate::{protocols::light_client::components::verify_extra_hash, storage::{HeaderWithExtension, db::StorageBatchRelatedOperations}};
 
 use super::{
     super::{LightClientProtocol, Status, StatusCode},
