+++ light-client-lib/src/storage/db/mod.rs
@@ -1,20 +1,22 @@
 #[cfg(all(not(target_arch = "wasm32"), feature = "rocksdb"))]
 mod native_rocksdb;
+use ckb_types::core::cell::CellStatus;
+use ckb_types::packed::OutPoint;
 #[cfg(all(not(target_arch = "wasm32"), feature = "rocksdb"))]
 pub use native_rocksdb::{Batch, Storage};
 
 #[cfg(all(not(target_arch = "wasm32"), feature = "rusqlite"))]
 mod native_rusqlite;
 #[cfg(all(not(target_arch = "wasm32"), feature = "rusqlite"))]
-pub use native_rusqlite::{Batch, CursorDirection, Storage, KV};
+pub use native_rusqlite::{Batch, Storage, KV};
 
 #[cfg(target_arch = "wasm32")]
 mod browser;
 #[cfg(target_arch = "wasm32")]
-pub use browser::{Batch, CursorDirection, Storage};
+pub use browser::{Batch, Storage};
 
-#[cfg(not(target_arch = "wasm32"))]
-use ckb_traits::HeaderProvider;
+
+use crate::error::Result;
 use ckb_types::prelude::Reader;
 use ckb_types::prelude::{Builder, FromSliceShouldBeOk};
 use ckb_types::{
@@ -34,11 +36,15 @@ use crate::storage::{
 use ckb_types::prelude::Entity;
 use ckb_types::U256;
 
-pub enum GetMatchedBlocksDirection {
+#[derive(Clone, Copy)]
+pub enum GeneralDirection {
     Forward,
     Reverse,
 }
 
+/**
+ * These functions will be implemented in each database implementation
+ */
 pub trait StorageHighLevelOperations {
     fn is_filter_scripts_empty(&self) -> bool;
     fn get_filter_scripts(&self) -> Vec<ScriptStatus>;
@@ -47,7 +53,7 @@ pub trait StorageHighLevelOperations {
     // get scripts hash that should be filtered below the given block number
     fn get_scripts_hash(&self, block_number: BlockNumber) -> Vec<Byte32>;
     fn clear_matched_blocks(&self);
-    fn get_matched_blocks(&self, direction: GetMatchedBlocksDirection) -> Option<MatchedBlocks>;
+    fn get_matched_blocks(&self, direction: GeneralDirection) -> Option<MatchedBlocks>;
     fn get_earliest_matched_blocks(&self) -> Option<MatchedBlocks>;
     fn get_latest_matched_blocks(&self) -> Option<MatchedBlocks>;
     fn get_check_points(&self, start_index: CpIndex, limit: usize) -> Vec<Byte32>;
@@ -56,10 +62,135 @@ pub trait StorageHighLevelOperations {
     ///
     /// N.B. The specified block will be removed.
     fn rollback_to_block(&self, to_number: BlockNumber);
+
+    fn collect_iterator(
+        &self,
+        start_key_bound: Vec<u8>,
+        order: GeneralDirection,
+        take_while: Box<dyn Fn(&[u8]) -> bool + Send + 'static>,
+        filter_map: Box<dyn Fn(&[u8]) -> Option<Vec<u8>> + Send + 'static>,
+        limit: usize,
+        skip: usize,
+    ) -> Vec<(Vec<u8>, Vec<u8>)>;
+    fn cell(&self, out_point: &OutPoint, eager_load: bool) -> CellStatus;
+    fn get<K: AsRef<[u8]>>(&self, key: K) -> Result<Option<Vec<u8>>>;
+    fn put<K, V>(&self, key: K, value: V) -> Result<()>
+    where
+        K: AsRef<[u8]>,
+        V: AsRef<[u8]>;
+    fn delete<K: AsRef<[u8]>>(&self, key: K) -> Result<()>;
+    fn get_header(&self, hash: &Byte32) -> Option<HeaderView>;
 }
 
-impl Storage {
-    pub fn init_genesis_block(&self, block: Block) {
+/**
+ * These functions can be automatically implemented based on StorageHighLevelOperations
+ */
+pub trait StorageGeneralOperations {
+    fn get_genesis_block(&self) -> Block;
+    fn update_last_state(
+        &self,
+        total_difficulty: &U256,
+        tip_header: &Header,
+        last_n_headers: &[HeaderView],
+    );
+    fn update_last_n_headers(&self, headers: &[HeaderView]);
+    /// 0 all blocks downloaded and inserted into storage call this function.
+    fn remove_matched_blocks(&self, start_number: u64);
+    /// the matched blocks must not empty
+    fn add_matched_blocks(
+        &self,
+        start_number: u64,
+        blocks_count: u64,
+        // (block-hash, proved)
+        matched_blocks: Vec<(Byte32, bool)>,
+    );
+    fn cleanup_invalid_matched_blocks(&self);
+
+    fn get_tip_header(&self) -> Header;
+    fn update_min_filtered_block_number(&self, block_number: BlockNumber);
+    fn get_last_check_point(&self) -> (CpIndex, Byte32);
+    fn update_max_check_point_index(&self, index: CpIndex);
+
+    fn get_transaction(&self, tx_hash: &Byte32) -> Option<(BlockNumber, TxIndex, Transaction)>;
+    fn get_transaction_with_header(&self, tx_hash: &Byte32) -> Option<(Transaction, Header)>;
+}
+
+/**
+ * These functions will call `Storage::get_pinned`, which has different return value type between RocksDB and other implementations
+ * This trait will be implemented in this module file.
+ */
+pub trait StorageGetPinnedRelatedOperations {
+    fn get_last_state(&self) -> (U256, Header);
+    fn get_last_n_headers(&self) -> Vec<(u64, Byte32)>;
+    fn get_min_filtered_block_number(&self) -> BlockNumber;
+    fn get_max_check_point_index(&self) -> CpIndex;
+}
+
+/**
+ * These functions will use `Batch`, which has different implementations among database implementations.
+ * This trait will be implemented in this module file.
+ */
+pub trait StorageBatchRelatedOperations {
+    fn init_genesis_block(&self, block: Block);
+    fn add_fetched_header(&self, hwe: &HeaderWithExtension);
+    fn add_fetched_tx(&self, tx: &Transaction, hwe: &HeaderWithExtension);
+    fn update_check_points(&self, start_index: CpIndex, check_points: &[Byte32]);
+    fn filter_block(&self, block: Block);
+}
+
+impl StorageGetPinnedRelatedOperations for Storage {
+    fn get_max_check_point_index(&self) -> CpIndex {
+        let key = Key::Meta(MAX_CHECK_POINT_INDEX).into_vec();
+        self.get_pinned(&key)
+            .expect("db get max check point index should be ok")
+            .map(|data| CpIndex::from_be_bytes(AsRef::<[u8]>::as_ref(&data).try_into().unwrap()))
+            .expect("db get max check point index should be ok")
+    }
+    fn get_min_filtered_block_number(&self) -> BlockNumber {
+        let key = Key::Meta(MIN_FILTERED_BLOCK_NUMBER).into_vec();
+        self.get_pinned(&key)
+            .expect("db get min filtered block number should be ok")
+            .map(|data| u64::from_le_bytes(AsRef::<[u8]>::as_ref(&data).try_into().unwrap()))
+            .unwrap_or_default()
+    }
+
+    fn get_last_n_headers(&self) -> Vec<(u64, Byte32)> {
+        let key = Key::Meta(LAST_N_HEADERS_KEY).into_vec();
+        self.get_pinned(&key)
+            .expect("db get last n headers should be ok")
+            .map(|data| {
+                assert!(AsRef::<[u8]>::as_ref(&data).len().is_multiple_of(40));
+                let mut headers = Vec::with_capacity(&AsRef::<[u8]>::as_ref(&data).len() / 40);
+                for part in AsRef::<[u8]>::as_ref(&data).chunks(40) {
+                    let number = u64::from_le_bytes(part[0..8].try_into().unwrap());
+                    let hash = Byte32::from_slice(&part[8..]).expect("byte32 block hash");
+                    headers.push((number, hash));
+                }
+                headers
+            })
+            .expect("last n headers should be inited")
+    }
+
+    fn get_last_state(&self) -> (U256, Header) {
+        let key = Key::Meta(LAST_STATE_KEY).into_vec();
+        self.get_pinned(&key)
+            .expect("db get last state should be ok")
+            .map(|data| {
+                let mut total_difficulty_bytes = [0u8; 32];
+                total_difficulty_bytes.copy_from_slice(&AsRef::<[u8]>::as_ref(&data)[0..32]);
+                let total_difficulty = U256::from_le_bytes(&total_difficulty_bytes);
+                let header = packed::HeaderReader::from_slice_should_be_ok(
+                    &AsRef::<[u8]>::as_ref(&data)[32..],
+                )
+                .to_entity();
+                (total_difficulty, header)
+            })
+            .expect("tip header should be inited")
+    }
+}
+
+impl StorageBatchRelatedOperations for Storage {
+    fn init_genesis_block(&self, block: Block) {
         let genesis_hash = block.calc_header_hash();
         let genesis_block_key = Key::Meta(GENESIS_BLOCK_KEY).into_vec();
         if let Some(stored_genesis_hash) = self
@@ -119,181 +250,7 @@ impl Storage {
             self.update_min_filtered_block_number(0);
         }
     }
-
-    pub fn get_genesis_block(&self) -> Block {
-        let genesis_hash_and_txs_hash = self
-            .get(Key::Meta(GENESIS_BLOCK_KEY).into_vec())
-            .expect("get genesis block")
-            .expect("inited storage");
-        let genesis_hash = Byte32::from_slice(&genesis_hash_and_txs_hash[0..32])
-            .expect("stored genesis block hash");
-        let genesis_header = Header::from_slice(
-            &self
-                .get(Key::BlockHash(&genesis_hash).into_vec())
-                .expect("db get should be ok")
-                .expect("stored block hash / header mapping"),
-        )
-        .expect("stored header should be OK");
-
-        let transactions: Vec<Transaction> = genesis_hash_and_txs_hash[32..]
-            .chunks_exact(32)
-            .map(|tx_hash| {
-                Transaction::from_slice(
-                    &self
-                        .get(
-                            Key::TxHash(
-                                &Byte32::from_slice(tx_hash).expect("stored genesis block tx hash"),
-                            )
-                            .into_vec(),
-                        )
-                        .expect("db get should be ok")
-                        .expect("stored genesis block tx")[12..],
-                )
-                .expect("stored Transaction")
-            })
-            .collect();
-
-        Block::new_builder()
-            .header(genesis_header)
-            .transactions(transactions.pack())
-            .build()
-    }
-
-    pub fn update_last_state(
-        &self,
-        total_difficulty: &U256,
-        tip_header: &Header,
-        last_n_headers: &[HeaderView],
-    ) {
-        let key = Key::Meta(LAST_STATE_KEY).into_vec();
-        let mut value = total_difficulty.to_le_bytes().to_vec();
-        value.extend(tip_header.as_slice());
-        self.put(key, &value)
-            .expect("db put last state should be ok");
-        self.update_last_n_headers(last_n_headers);
-    }
-
-    pub fn get_last_state(&self) -> (U256, Header) {
-        let key = Key::Meta(LAST_STATE_KEY).into_vec();
-        self.get_pinned(&key)
-            .expect("db get last state should be ok")
-            .map(|data| {
-                let mut total_difficulty_bytes = [0u8; 32];
-                total_difficulty_bytes.copy_from_slice(&AsRef::<[u8]>::as_ref(&data)[0..32]);
-                let total_difficulty = U256::from_le_bytes(&total_difficulty_bytes);
-                let header = packed::HeaderReader::from_slice_should_be_ok(
-                    &AsRef::<[u8]>::as_ref(&data)[32..],
-                )
-                .to_entity();
-                (total_difficulty, header)
-            })
-            .expect("tip header should be inited")
-    }
-
-    fn update_last_n_headers(&self, headers: &[HeaderView]) {
-        let key = Key::Meta(LAST_N_HEADERS_KEY).into_vec();
-        let mut value: Vec<u8> = Vec::with_capacity(headers.len() * 40);
-        for header in headers {
-            value.extend(header.number().to_le_bytes());
-            value.extend(header.hash().as_slice());
-        }
-        self.put(key, &value)
-            .expect("db put last n headers should be ok");
-    }
-
-    pub fn get_last_n_headers(&self) -> Vec<(u64, Byte32)> {
-        let key = Key::Meta(LAST_N_HEADERS_KEY).into_vec();
-        self.get_pinned(&key)
-            .expect("db get last n headers should be ok")
-            .map(|data| {
-                assert!(AsRef::<[u8]>::as_ref(&data).len().is_multiple_of(40));
-                let mut headers = Vec::with_capacity(&AsRef::<[u8]>::as_ref(&data).len() / 40);
-                for part in AsRef::<[u8]>::as_ref(&data).chunks(40) {
-                    let number = u64::from_le_bytes(part[0..8].try_into().unwrap());
-                    let hash = Byte32::from_slice(&part[8..]).expect("byte32 block hash");
-                    headers.push((number, hash));
-                }
-                headers
-            })
-            .expect("last n headers should be inited")
-    }
-
-    /// 0 all blocks downloaded and inserted into storage call this function.
-    pub fn remove_matched_blocks(&self, start_number: u64) {
-        let mut key = Key::Meta(MATCHED_FILTER_BLOCKS_KEY).into_vec();
-        key.extend(start_number.to_be_bytes());
-        self.delete(&key).expect("delete matched blocks");
-    }
-
-    /// the matched blocks must not empty
-    pub fn add_matched_blocks(
-        &self,
-        start_number: u64,
-        blocks_count: u64,
-        // (block-hash, proved)
-        matched_blocks: Vec<(Byte32, bool)>,
-    ) {
-        assert!(!matched_blocks.is_empty());
-        let mut key = Key::Meta(MATCHED_FILTER_BLOCKS_KEY).into_vec();
-        key.extend(start_number.to_be_bytes());
-
-        let mut value = blocks_count.to_le_bytes().to_vec();
-        for (block_hash, proved) in matched_blocks {
-            value.extend(block_hash.as_slice());
-            value.push(u8::from(proved));
-        }
-        self.put(key, &value)
-            .expect("db put matched blocks should be ok");
-    }
-
-    pub fn cleanup_invalid_matched_blocks(&self) {
-        use ckb_types::prelude::Unpack;
-        use log::warn;
-
-        let tip_number: u64 = self.get_tip_header().raw().number().unpack();
-
-        loop {
-            let entry = self.get_earliest_matched_blocks();
-            if entry.is_none() {
-                break;
-            }
-
-            let matched_blocks = entry.unwrap();
-            let start_number = matched_blocks.start_number;
-            let blocks_count = matched_blocks.blocks_count;
-            let mut should_remove = false;
-
-            for block in &matched_blocks.blocks {
-                if let Some(header) = self.get_header(&block.hash) {
-                    let stored_number: u64 = header.number();
-                    if stored_number < start_number || stored_number >= start_number + blocks_count
-                    {
-                        warn!(
-                            "Invalid matched block {:#x} at number {} outside expected range [{}, {}), removing entry at start_number={}",
-                            block.hash, stored_number, start_number, start_number + blocks_count, start_number
-                        );
-                        should_remove = true;
-                        break;
-                    }
-                } else if start_number + 1000 < tip_number {
-                    warn!(
-                        "Matched block {:#x} not found in storage, entry at start_number={} is {} blocks behind tip, removing",
-                        block.hash, start_number, tip_number - start_number
-                    );
-                    should_remove = true;
-                    break;
-                }
-            }
-
-            if should_remove {
-                self.remove_matched_blocks(start_number);
-            } else {
-                break;
-            }
-        }
-    }
-
-    pub fn add_fetched_header(&self, hwe: &HeaderWithExtension) {
+    fn add_fetched_header(&self, hwe: &HeaderWithExtension) {
         let mut batch = self.batch();
         let block_hash = hwe.header.calc_header_hash();
         batch
@@ -308,7 +265,7 @@ impl Storage {
         batch.commit().expect("batch commit should be ok");
     }
 
-    pub fn add_fetched_tx(&self, tx: &Transaction, hwe: &HeaderWithExtension) {
+    fn add_fetched_tx(&self, tx: &Transaction, hwe: &HeaderWithExtension) {
         let mut batch = self.batch();
         let block_hash = hwe.header.calc_header_hash();
         let block_number: u64 = hwe.header.raw().number().unpack();
@@ -328,52 +285,7 @@ impl Storage {
         batch.put_kv(key, value).expect("batch put should be ok");
         batch.commit().expect("batch commit should be ok");
     }
-
-    pub fn get_tip_header(&self) -> Header {
-        self.get_last_state().1
-    }
-
-    pub fn get_min_filtered_block_number(&self) -> BlockNumber {
-        let key = Key::Meta(MIN_FILTERED_BLOCK_NUMBER).into_vec();
-        self.get_pinned(&key)
-            .expect("db get min filtered block number should be ok")
-            .map(|data| u64::from_le_bytes(AsRef::<[u8]>::as_ref(&data).try_into().unwrap()))
-            .unwrap_or_default()
-    }
-
-    pub fn update_min_filtered_block_number(&self, block_number: BlockNumber) {
-        let key = Key::Meta(MIN_FILTERED_BLOCK_NUMBER).into_vec();
-        let value = block_number.to_le_bytes();
-        self.put(key, value)
-            .expect("db put min filtered block number should be ok");
-    }
-
-    pub fn get_last_check_point(&self) -> (CpIndex, Byte32) {
-        let index = self.get_max_check_point_index();
-        let hash = self
-            .get_check_points(index, 1)
-            .first()
-            .cloned()
-            .expect("db get last check point should be ok");
-        (index, hash)
-    }
-
-    pub fn get_max_check_point_index(&self) -> CpIndex {
-        let key = Key::Meta(MAX_CHECK_POINT_INDEX).into_vec();
-        self.get_pinned(&key)
-            .expect("db get max check point index should be ok")
-            .map(|data| CpIndex::from_be_bytes(AsRef::<[u8]>::as_ref(&data).try_into().unwrap()))
-            .expect("db get max check point index should be ok")
-    }
-
-    pub fn update_max_check_point_index(&self, index: CpIndex) {
-        let key = Key::Meta(MAX_CHECK_POINT_INDEX).into_vec();
-        let value = index.to_be_bytes();
-        self.put(key, value)
-            .expect("db put max check point index should be ok");
-    }
-
-    pub fn update_check_points(&self, start_index: CpIndex, check_points: &[Byte32]) {
+    fn update_check_points(&self, start_index: CpIndex, check_points: &[Byte32]) {
         let mut index = start_index;
         let mut batch = self.batch();
         for cp in check_points {
@@ -385,7 +297,7 @@ impl Storage {
         batch.commit().expect("batch commit should be ok");
     }
 
-    pub fn filter_block(&self, block: Block) {
+    fn filter_block(&self, block: Block) {
         let scripts: HashSet<(Script, ScriptType)> = self
             .get_filter_scripts()
             .into_iter()
@@ -581,6 +493,177 @@ impl Storage {
         }
         batch.commit().expect("batch commit should be ok");
     }
+}
+
+impl<T: StorageHighLevelOperations + StorageGetPinnedRelatedOperations> StorageGeneralOperations
+    for T
+{
+    fn get_genesis_block(&self) -> Block {
+        let genesis_hash_and_txs_hash = self
+            .get(Key::Meta(GENESIS_BLOCK_KEY).into_vec())
+            .expect("get genesis block")
+            .expect("inited storage");
+        let genesis_hash = Byte32::from_slice(&genesis_hash_and_txs_hash[0..32])
+            .expect("stored genesis block hash");
+        let genesis_header = Header::from_slice(
+            &self
+                .get(Key::BlockHash(&genesis_hash).into_vec())
+                .expect("db get should be ok")
+                .expect("stored block hash / header mapping"),
+        )
+        .expect("stored header should be OK");
+
+        let transactions: Vec<Transaction> = genesis_hash_and_txs_hash[32..]
+            .chunks_exact(32)
+            .map(|tx_hash| {
+                Transaction::from_slice(
+                    &self
+                        .get(
+                            Key::TxHash(
+                                &Byte32::from_slice(tx_hash).expect("stored genesis block tx hash"),
+                            )
+                            .into_vec(),
+                        )
+                        .expect("db get should be ok")
+                        .expect("stored genesis block tx")[12..],
+                )
+                .expect("stored Transaction")
+            })
+            .collect();
+
+        Block::new_builder()
+            .header(genesis_header)
+            .transactions(transactions.pack())
+            .build()
+    }
+
+    fn update_last_state(
+        &self,
+        total_difficulty: &U256,
+        tip_header: &Header,
+        last_n_headers: &[HeaderView],
+    ) {
+        let key = Key::Meta(LAST_STATE_KEY).into_vec();
+        let mut value = total_difficulty.to_le_bytes().to_vec();
+        value.extend(tip_header.as_slice());
+        self.put(key, &value)
+            .expect("db put last state should be ok");
+        self.update_last_n_headers(last_n_headers);
+    }
+
+    fn update_last_n_headers(&self, headers: &[HeaderView]) {
+        let key = Key::Meta(LAST_N_HEADERS_KEY).into_vec();
+        let mut value: Vec<u8> = Vec::with_capacity(headers.len() * 40);
+        for header in headers {
+            value.extend(header.number().to_le_bytes());
+            value.extend(header.hash().as_slice());
+        }
+        self.put(key, &value)
+            .expect("db put last n headers should be ok");
+    }
+
+    /// 0 all blocks downloaded and inserted into storage call this function.
+    fn remove_matched_blocks(&self, start_number: u64) {
+        let mut key = Key::Meta(MATCHED_FILTER_BLOCKS_KEY).into_vec();
+        key.extend(start_number.to_be_bytes());
+        self.delete(&key).expect("delete matched blocks");
+    }
+
+    /// the matched blocks must not empty
+    fn add_matched_blocks(
+        &self,
+        start_number: u64,
+        blocks_count: u64,
+        // (block-hash, proved)
+        matched_blocks: Vec<(Byte32, bool)>,
+    ) {
+        assert!(!matched_blocks.is_empty());
+        let mut key = Key::Meta(MATCHED_FILTER_BLOCKS_KEY).into_vec();
+        key.extend(start_number.to_be_bytes());
+
+        let mut value = blocks_count.to_le_bytes().to_vec();
+        for (block_hash, proved) in matched_blocks {
+            value.extend(block_hash.as_slice());
+            value.push(u8::from(proved));
+        }
+        self.put(key, &value)
+            .expect("db put matched blocks should be ok");
+    }
+
+    fn cleanup_invalid_matched_blocks(&self) {
+        use ckb_types::prelude::Unpack;
+        use log::warn;
+
+        let tip_number: u64 = self.get_tip_header().raw().number().unpack();
+
+        loop {
+            let entry = self.get_earliest_matched_blocks();
+            if entry.is_none() {
+                break;
+            }
+
+            let matched_blocks = entry.unwrap();
+            let start_number = matched_blocks.start_number;
+            let blocks_count = matched_blocks.blocks_count;
+            let mut should_remove = false;
+
+            for block in &matched_blocks.blocks {
+                if let Some(header) = self.get_header(&block.hash) {
+                    let stored_number: u64 = header.number();
+                    if stored_number < start_number || stored_number >= start_number + blocks_count
+                    {
+                        warn!(
+                            "Invalid matched block {:#x} at number {} outside expected range [{}, {}), removing entry at start_number={}",
+                            block.hash, stored_number, start_number, start_number + blocks_count, start_number
+                        );
+                        should_remove = true;
+                        break;
+                    }
+                } else if start_number + 1000 < tip_number {
+                    warn!(
+                        "Matched block {:#x} not found in storage, entry at start_number={} is {} blocks behind tip, removing",
+                        block.hash, start_number, tip_number - start_number
+                    );
+                    should_remove = true;
+                    break;
+                }
+            }
+
+            if should_remove {
+                self.remove_matched_blocks(start_number);
+            } else {
+                break;
+            }
+        }
+    }
+
+    fn get_tip_header(&self) -> Header {
+        self.get_last_state().1
+    }
+
+    fn update_min_filtered_block_number(&self, block_number: BlockNumber) {
+        let key = Key::Meta(MIN_FILTERED_BLOCK_NUMBER).into_vec();
+        let value = block_number.to_le_bytes();
+        self.put(key, value)
+            .expect("db put min filtered block number should be ok");
+    }
+
+    fn get_last_check_point(&self) -> (CpIndex, Byte32) {
+        let index = self.get_max_check_point_index();
+        let hash = self
+            .get_check_points(index, 1)
+            .first()
+            .cloned()
+            .expect("db get last check point should be ok");
+        (index, hash)
+    }
+
+    fn update_max_check_point_index(&self, index: CpIndex) {
+        let key = Key::Meta(MAX_CHECK_POINT_INDEX).into_vec();
+        let value = index.to_be_bytes();
+        self.put(key, value)
+            .expect("db put max check point index should be ok");
+    }
 
     fn get_transaction(&self, tx_hash: &Byte32) -> Option<(BlockNumber, TxIndex, Transaction)> {
         self.get(Key::TxHash(tx_hash).into_vec())
@@ -596,7 +679,7 @@ impl Storage {
             .expect("db get should be ok")
     }
 
-    pub fn get_transaction_with_header(&self, tx_hash: &Byte32) -> Option<(Transaction, Header)> {
+    fn get_transaction_with_header(&self, tx_hash: &Byte32) -> Option<(Transaction, Header)> {
         self.get_transaction(tx_hash)
             .map(|(block_number, _tx_index, tx)| {
                 let block_hash = Byte32::from_slice(
