+++ light-client-lib/src/storage/db/native_rusqlite.rs
@@ -3,10 +3,12 @@ use super::super::{
     KeyPrefix, MatchedBlock, MatchedBlocks, Script, FILTER_SCRIPTS_KEY, MATCHED_FILTER_BLOCKS_KEY,
     MIN_FILTERED_BLOCK_NUMBER,
 };
+use crate::storage::db::StorageBatchRelatedOperations;
+use crate::storage::db::StorageGetPinnedRelatedOperations;
 use crate::{
     error::Result,
     storage::{
-        db::{GetMatchedBlocksDirection, StorageHighLevelOperations},
+        db::{GeneralDirection, StorageGeneralOperations, StorageHighLevelOperations},
         ScriptStatus, ScriptType, SetScriptsCommand, TxIndex,
     },
 };
@@ -179,6 +181,53 @@ impl Storage {
 }
 
 impl StorageHighLevelOperations for Storage {
+    fn get_header(&self, hash: &Byte32) -> Option<HeaderView> {
+        HeaderProvider::get_header(self, hash)
+    }
+    fn put<K, V>(&self, key: K, value: V) -> Result<()>
+    where
+        K: AsRef<[u8]>,
+        V: AsRef<[u8]>,
+    {
+        self.put(key, value)
+    }
+
+    fn delete<K: AsRef<[u8]>>(&self, key: K) -> Result<()> {
+        self.delete(key)
+    }
+    fn collect_iterator(
+        &self,
+        start_key_bound: Vec<u8>,
+        order: GeneralDirection,
+        take_while: Box<dyn Fn(&[u8]) -> bool + Send + 'static>,
+        filter_map: Box<dyn Fn(&[u8]) -> Option<Vec<u8>> + Send + 'static>,
+        limit: usize,
+        skip: usize,
+    ) -> Vec<(Vec<u8>, Vec<u8>)> {
+        self.collect_iterator(
+            &start_key_bound,
+            match order {
+                GeneralDirection::Forward => CursorDirection::Ascending,
+                GeneralDirection::Reverse => CursorDirection::Descending,
+            },
+            take_while,
+            |key, value| {
+                filter_map(key).map(|key| KV {
+                    key: key,
+                    value: value.to_vec(),
+                })
+            },
+            limit,
+            skip,
+        )
+        .unwrap()
+        .into_iter()
+        .map(|x| (x.key, x.value))
+        .collect()
+    }
+    fn get<K: AsRef<[u8]>>(&self, key: K) -> Result<Option<Vec<u8>>> {
+        self.get(key)
+    }
     fn is_filter_scripts_empty(&self) -> bool {
         let key_prefix = Key::Meta(FILTER_SCRIPTS_KEY).into_vec();
         self.collect_iterator(
@@ -440,11 +489,11 @@ impl StorageHighLevelOperations for Storage {
         batch.commit().unwrap();
     }
 
-    fn get_matched_blocks(&self, direction: GetMatchedBlocksDirection) -> Option<MatchedBlocks> {
+    fn get_matched_blocks(&self, direction: GeneralDirection) -> Option<MatchedBlocks> {
         let key_prefix = Key::Meta(MATCHED_FILTER_BLOCKS_KEY).into_vec();
         let iter_from = match direction {
-            GetMatchedBlocksDirection::Forward => key_prefix.clone(),
-            GetMatchedBlocksDirection::Reverse => {
+            GeneralDirection::Forward => key_prefix.clone(),
+            GeneralDirection::Reverse => {
                 let mut key = key_prefix.clone();
                 key.extend(u64::MAX.to_be_bytes());
                 key
@@ -455,8 +504,8 @@ impl StorageHighLevelOperations for Storage {
             .collect_iterator(
                 &iter_from,
                 match direction {
-                    GetMatchedBlocksDirection::Forward => CursorDirection::Ascending,
-                    GetMatchedBlocksDirection::Reverse => CursorDirection::Descending,
+                    GeneralDirection::Forward => CursorDirection::Ascending,
+                    GeneralDirection::Reverse => CursorDirection::Descending,
                 },
                 |x| x.starts_with(&key_prefix),
                 |k, v| {
@@ -492,11 +541,11 @@ impl StorageHighLevelOperations for Storage {
     }
 
     fn get_earliest_matched_blocks(&self) -> Option<MatchedBlocks> {
-        self.get_matched_blocks(GetMatchedBlocksDirection::Forward)
+        self.get_matched_blocks(GeneralDirection::Forward)
     }
 
     fn get_latest_matched_blocks(&self) -> Option<MatchedBlocks> {
-        self.get_matched_blocks(GetMatchedBlocksDirection::Reverse)
+        self.get_matched_blocks(GeneralDirection::Reverse)
     }
 
     fn get_check_points(&self, start_index: CpIndex, limit: usize) -> Vec<Byte32> {
@@ -728,6 +777,9 @@ impl StorageHighLevelOperations for Storage {
         }
 
         batch.commit().expect("batch commit should be ok");
+    }
+    fn cell(&self, out_point: &OutPoint, eager_load: bool) -> CellStatus {
+        CellProvider::cell(self, out_point, eager_load)
     }
 }
