+++ light-client-lib/src/tests/protocols/light_client/send_last_state_proof.rs
@@ -6,7 +6,7 @@ use ckb_types::{
     H256, U256,
 };
 use log::debug;
-use crate::storage::db::StorageHighLevelOperations;
+use crate::storage::db::{StorageGeneralOperations, StorageGetPinnedRelatedOperations, StorageHighLevelOperations};
 use crate::{
     protocols::{light_client::prelude::*, LastState, ProveRequest, ProveState, StatusCode},
     tests::{
